refactor : flower 엔티티 삭제 및 구조 리팩토링
