--- src/main/resources/templates/login.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>로그인 페이지</title>
-</head>
-<body>
-<h1>로그인 페이지</h1>
-<hr/>
-<!-- 시큐리티는 x-www-form-url-encoded 타입만 인식 -->
-<form action="/loginProc" method="post">
-	<input type="text" name="username" />
-	<input type="password" name="password" />
-	<button>로그인</button>
-</form>
-
-<br />
-
-	<h1>Social Login</h1>
-	<br />
-	<!-- javascript:; 는 클릭해도 반응을 없게 하는 키워드 -->
-<!--	<a href="/oauth2/authorization/google" >-->
-<!--	<img src="https://pngimage.net/wp-content/uploads/2018/06/google-login-button-png-1.png"-->
-<!--		alt="google" width="357px" height="117px">-->
-<!--	</a>-->
-	<br />
-	<a href="/oauth2/authorization/naver">
-	<img src="https://blogfiles.pstatic.net/MjAyMDA4MDRfMzMg/MDAxNTk2NTEyOTY4MDMx.vhXHCulffijGUnvlaBR2jW4__Lkz8R3ZTaEDcTeNV2gg.Wt_HNl_zktUJUMrYGkVmqJ-PhxKv_s4A7gG1uPKMZaQg.PNG.getinthere/naver_button.png"
-		alt="naver" width="357px" height="50px">
-	</a>
-	<br />
-	<a href="/oauth2/authorization/kakao">
-		<img src="https://blogfiles.pstatic.net/MjAyMDA4MDRfMzMg/MDAxNTk2NTEyOTY4MDMx.vhXHCulffijGUnvlaBR2jW4__Lkz8R3ZTaEDcTeNV2gg.Wt_HNl_zktUJUMrYGkVmqJ-PhxKv_s4A7gG1uPKMZaQg.PNG.getinthere/naver_button.png"
-			 alt="kakao" width="357px" height="50px">
-	</a>
-	<br />
-</body>
-</html>
